Create new CPU class, move a bunch of stuff
