+++ nops_proj.csproj
@@ -186,6 +186,7 @@
   <ItemGroup>
     <Compile Include="Bridge.cs" />
     <Compile Include="GDBServer.cs" />
+    <Compile Include="CPU.cs" />
     <Compile Include="TargetDataPort.cs" />
     <Compile Include="SerialTarget.cs" />
     <Compile Include="NOTPSXSERIAL.CS" />
